+++ agentinsight/langchain/__init__.py
@@ -1,5 +1,7 @@
 from .CallbackHandler import (
     LangchainCallbackHandler as CallbackHandler,
-)  # For backward compatibility
+)
 
-__all__ = ["CallbackHandler"]
+AgentInsightCallbackHandler = CallbackHandler  # For backward compatibility
+
+__all__ = ["CallbackHandler", "AgentInsightCallbackHandler"]